Add package.json with ws dependency
Having a package.json will allow all the dependencies to be installed very
easily simply by the user running "npm install" after obtaining the source
code. Also, when installing the dependencies this way, they get installed
locally (in the node_modules subdirectory in the source tree) so there is
no need to have root privilege to install them. Last but not least this could
be used to push releases to npm, so that the package can be installed on a
machine just using "npm install -g dewdle".